+++ tests/keyspace/test_keyspace.py
@@ -6,8 +6,12 @@ import rembus
 
 # def consume(ctx, rb, topic, data):
 def consume(topic, data, ctx, node):
-    logging.info("[%s] recv: %s", topic, data)
+    logging.info("[%s] test_keyspace recv: %s", topic, data)
     ctx[node.rid] = data
+
+
+def broker_consume(topic, data):
+    logging.info("[%s] broker recv: %s", topic, data)
 
 
 async def create_sub(name, topic, ctx):
@@ -20,18 +24,22 @@ async def create_sub(name, topic, ctx):
 
 @pytest.mark.anyio
 async def test_add_plugin():
+    space_topic = "k/*/y"
     ctx = {}
     server = await rembus.component(port=8000)
 
+    # register a space topic to the server
+    await server.subscribe(broker_consume, topic=space_topic)
+
     pub = await rembus.component("pub")
 
-    sub1 = await create_sub("sub1", "k/*/y", ctx)
-    sub2 = await create_sub("sub2", "k/*/y", ctx)
+    sub1 = await create_sub("sub1", space_topic, ctx)
+    sub2 = await create_sub("sub2", space_topic, ctx)
 
     await pub.publish("k/e/y", "space")
     await pub.publish("k/@e/y", "space")
 
-    await sub2.unsubscribe("k/*/y")
+    await sub2.unsubscribe(space_topic)
     await sub1.close()
 
     await asyncio.sleep(0.1)
@@ -44,3 +52,33 @@ async def test_add_plugin():
     await server.close()
 
     assert len(ctx) == 2
+
+
+@pytest.mark.anyio
+async def test_remove_close_from_keyspace():
+    space_topic = "k/*/y"
+    ctx = {}
+    server = await rembus.component(port=8000)
+
+    pub = await rembus.component("pub")
+
+    sub1 = await create_sub("sub1", space_topic, ctx)
+    sub2 = await create_sub("sub2", space_topic, ctx)
+
+    await pub.publish("k/e/y", "space")
+
+    await asyncio.sleep(0.1)
+    await sub1.close()
+
+    # sub1 twin is gone, deliver only to sub2.
+    await asyncio.sleep(0.1)
+    await pub.publish("k/e/y", "space_1999")
+
+    await asyncio.sleep(0.1)
+    await pub.close()
+    await sub2.close()
+    await server.close()
+
+    assert len(ctx) == 2
+    assert ctx["sub1"] == "space"
+    assert ctx["sub2"] == "space_1999"